+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Reflect API changes:
   - Now `Transaction` has `attachments: List<Attachment>` and `labels: List<Label>` fields - `attachmentIds`
     and `labelIds` are therefore now deprecated
-  - A new `getTransaction(id)` API to get a single `Transaction`
+  - A new `getTransaction(id)` API to get a single `Transaction` by id
 - Add a `bypassSslChecks` option on `HttpConfiguration` (should never be used in production)
 - Update all dependencies to their latest version